requirements.txt
aiogram>=3.7.0,<4.0.0
python-dotenv==1.0.1
aiosqlite==0.20.0
tzdata>=2024.1; platform_system == "Windows"
